Starting over - kind of
THIS IS STILL A WORK IN PROGRESS

I'm not satisfied with my setup. It doesn't seem like I'm using
lazy.nvim correctly, and there are files all over the place. This branch
will be an effort to taking what I've learned and using it to rebuild my
setup from scratch